Update scenarios:  * removed redundant steps  * optimized generics steps
