+++ src/test/resources/features/AddProductToCart.feature
@@ -7,27 +7,24 @@ Feature: Add product to Shopping cart
     And user clicks on 'cell-phones' Button and move to 'SubCategory' page
     And user clicks on 'htc-one-mini-blue' Button and move to 'Product' page
     And user press on 'addToCart' Button
+    And alert message is displayed
+    And alert message is closed
 
   Scenario: Add product to the Shopping cart
-    When alert message is displayed
-    And alert message is closed
-    And user press on 'shoppingCart' Button
+    When user press on 'shoppingCart' Button
     And 'ShoppingCart' page is displayed
     Then 'HTC One Mini Blue' product is displayed
 
   Scenario: Change product quantity from the Shopping cart
-    When alert message is displayed
-    And alert message is closed
-    And user press on 'shoppingCart' Button
+    When user press on 'shoppingCart' Button
     And 'ShoppingCart' page is displayed
-    And user fills in '2'
+    And user changes 'quantity' Field with following values:
+      | quantityField | 2 |
     And user press on 'updateShoppingCart' Button
-    Then product quantity changed correct
+    Then 'quantity' Field successfully updated
 
   Scenario: Price validation for multiple products
-    When alert message is displayed
-    And alert message is closed
-    And user press on 'addToCart' Button
+    When user press on 'addToCart' Button
     And alert message is displayed
     And alert message is closed
     And user press on 'shoppingCart' Button
